// front-end/src/api/models/barbeiroPublicoDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * API Agendador de Barbearia
 * Contrato OpenAPI da API REST do Agendador de Barbearia.
 * OpenAPI spec version: 1.0
 */
import type { FilialPublicaResumoDto } from './filialPublicaResumoDto';

export interface BarbeiroPublicoDto {
  id: number;
  idFilial: number;
  nomeProfissional: string;
  /** @nullable */
  descricao?: string | null;
  /** @nullable */
  fotoUrl?: string | null;
  filial: FilialPublicaResumoDto;
}