// front-end/src/api/models/filialPublicaDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * API Agendador de Barbearia
 * Contrato OpenAPI da API REST do Agendador de Barbearia.
 * OpenAPI spec version: 1.0
 */
import type { EnderecoPublicoDto } from './enderecoPublicoDto';

export interface FilialPublicaDto {
  id: number;
  nome: string;
  /** @nullable */
  telefone?: string | null;
  endereco: EnderecoPublicoDto;
}