// front-end/src/api/models/servicoPublicoDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * API Agendador de Barbearia
 * Contrato OpenAPI da API REST do Agendador de Barbearia.
 * OpenAPI spec version: 1.0
 */

export interface ServicoPublicoDto {
  id: number;
  nome: string;
  /** @nullable */
  descricao?: string | null;
  precoBase: string;
  duracaoMinutos: number;
}